Gates and Policies

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -7,6 +7,7 @@ use App\Models\Task;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -15,6 +16,8 @@ class TaskController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Task::class);
+
         return TaskResource::collection(auth()->user()->tasks()->get());
     }
 
@@ -23,6 +26,10 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        if (request()->user()->cannot('create', Task::class)) {
+            abort(403, "This action is not authorized.");
+        }
+
         $task = $request->user()->tasks()->create($request->validated());
 
         return TaskResource::make($task);
@@ -33,6 +40,8 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
+        Gate::authorize('view', $task);
+
         return TaskResource::make($task);
     }
 
@@ -41,6 +50,10 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        if ($request->user()->cannot('update', $task)) {
+            abort(403, "This action is not authorized.");
+        }
+
         $task->update($request->validated());
 
         return TaskResource::make($task);
@@ -51,6 +64,10 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        if (request()->user()->cannot('delete', $task)) {
+            abort(403, "This action is not authorized.");
+        }
+
         $task->delete();
 
         return response()->noContent();
